Move engine channel-count constants and StemChannel to audio/
Part of a broader effort to clean up include dependencies as a step
towards a modular build.

Created `audio/enginechannelcount.h` containing:
  - `kEngineChannelOutputCount`, `kMaxEngineChannelInputCount`,
    `kMaxSupportedStems` (signal-path channel-count constants)
  - `StemChannel` enum and `StemChannelSelection` flags type
    (previously in `engine/engine.h` under `#ifdef __STEM__`)

These are properties of the audio signal path, not of the engine
subsystem, so audio/ is the correct home.

Removed the duplicate definitions from `engine/engine.h`, which now
includes `audio/enginechannelcount.h` to re-export the symbols.

Replaced `#include "engine/engine.h"` with
`#include "audio/enginechannelcount.h"` in:
  - `audio/frame.h` (added direct `audio/signalinfo.h` which it was
    previously getting transitively)
  - `analyzer/constants.h`

This breaks the waveform/ → engine/ and analyzer/ → engine/ include
paths, allowing those modules to depend only on audio/ instead of the
full engine/ subtree.
