+++ src/engine/engine.h
@@ -1,29 +1,9 @@
 #pragma once
 
+#include "audio/enginechannelcount.h"
 #include "audio/signalinfo.h"
 
-
 namespace mixxx {
-static constexpr audio::ChannelCount kEngineChannelOutputCount =
-        audio::ChannelCount::stereo();
-static constexpr audio::ChannelCount kMaxEngineChannelInputCount =
-        audio::ChannelCount::stem();
-// The following constant is always defined as it used for the waveform data
-// struct, which must stay consistent, whether the STEM feature is enabled or
-// not.
-constexpr int kMaxSupportedStems = 4;
-#ifdef __STEM__
-enum class StemChannel {
-    First = 0x1,
-    Second = 0x2,
-    Third = 0x4,
-    Fourth = 0x8,
-
-    None = 0,
-    All = First | Second | Third | Fourth
-};
-Q_DECLARE_FLAGS(StemChannelSelection, StemChannel);
-#endif
 
 // Contains the information needed to process a buffer of audio
 class EngineParameters final {
